Create phonegeneratorfile.php
This file is a master first draft code use to generate leads via whatsapp by extracting all possible phone numbers within a country and save in a database for further testing if those numbers are on Telegram, Whatsapp amongst others.
To implement the code, you need to follow these steps:

Create a MySQL database with a table named "phone_numbers" that has one column named "number".

Create a PHP file and copy the complete code into it.

Replace the database connection details with your own. Edit the following lines in the code:
$servername = "localhost";
$username = "yourusername";
$password = "yourpassword";
$dbname = "phone_numbers_db";

Replace "yourusername" and "yourpassword" with your MySQL username and password respectively, and "phone_numbers_db" with the name of your database.

Upload the PHP file to your web server.

Create a HTML form that allows users to input the quantity of phone numbers to generate. The form should submit to the PHP file you created in step 4.

When the user submits the form, the PHP code will generate the specified number of phone numbers using the specified prefixes and add them to the database.

The PHP code will then display the generated phone numbers in a list.

Note: Before implementing the code, ensure that you have the necessary web server software and a MySQL database set up on your system.